+++ frontend/src/components/file_browser/file_explorer/FileIcon.tsx
@@ -5,7 +5,17 @@ import { ReactComponent as OpenFolderIcon } from "./img/openFolder.svg";
 import { ReactComponent as BaseFileIcon } from "./img/file.svg";
 import { ReactComponent as ActionFileIcon } from "./img/file-action.svg";
 
-function FileIcon({ is_dir, is_collapsed, name, group }) {
+function FileIcon({
+  is_dir,
+  is_collapsed,
+  name,
+  group,
+}: {
+  is_dir: boolean;
+  is_collapsed: boolean;
+  name: string;
+  group: string;
+}) {
   var returnVal = (
     <>
       <BaseFileIcon className="bt-arrow-icon" fill={"var(--icon)"} />
